+++ Changelog.md
@@ -4,7 +4,8 @@
 - Detect cores on nRF53 with readback protection.
 
 ### Fixed
-- Remove readback protection on nRF53
+- Remove readback protection on nRF53.
+- Dropping several hex files.
 
 ## 2.2.0 - 2022-01-13
 ### Added